requirements.txt
ai2-olmo
bm25s
datasets==2.18.0
evaluate==0.4.2
git+https://github.com/allenai/catwalk
ipdb
ipython
matplotlib
nltk==3.8.1
notebook
numpy
openai
PyYAML
scikit-learn
scipy
seaborn
spacy==3.7.4
torch
tqdm
transformers==4.47.0
